requirements.txt
numpy>=1.26.0
pandas>=2.2.0
scikit-learn>=1.4.0
xgboost>=2.0.0
matplotlib>=3.8.0
seaborn>=0.13.0
econml>=0.15.0
causalml>=0.15.0
pyyaml>=6.0
scipy>=1.12.0
joblib>=1.3.0
tqdm>=4.66.0
jupyter>=1.0.0
ipykernel>=6.29.0
